AclAuthorizationWrapperTest
Adapted from the old AclAuthorizationProcessorTest
